+++ championship_manager.go
@@ -24,7 +24,6 @@ import (
 	"github.com/heindl/caldav-go/icalendar/components"
 	"github.com/mitchellh/go-wordwrap"
 	"github.com/sirupsen/logrus"
-	"github.com/yuin/gopher-lua"
 )
 
 type ChampionshipManager struct {
@@ -498,10 +497,12 @@ func (cm *ChampionshipManager) StartEvent(championshipID string, eventID string,
 
 	event.RaceSetup.Cars = strings.Join(championship.ValidCarIDs(), ";")
 
-	err, event = championshipEventStartPlugin(event, championship)
+	if config.Lua.Enabled {
+		err = championshipEventStartPlugin(event, championship)
 
-	if err != nil {
-		logrus.WithError(err).Error("championship event start plugin script failed")
+		if err != nil {
+			logrus.WithError(err).Error("championship event start plugin script failed")
+		}
 	}
 
 	entryList := event.CombineEntryLists(championship)
@@ -587,69 +588,27 @@ func (cm *ChampionshipManager) StartEvent(championshipID string, eventID string,
 	}
 }
 
-func championshipEventStartPlugin(event *ChampionshipEvent, championship *Championship) (error, *ChampionshipEvent) {
-	l := lua.NewState()
-	defer l.Close()
-
-	// @TODO why can't I set the lua path properly using either of these methods?
-	// @TODO export LUA_PATH=$PWD/cmd/server-manager/plugins/?.lua in terminal makes it work
-	/*err := os.Setenv("LUA_PATH", "./plugins/?.lua")
-
-	if err != nil {
-		return err, event
-	}*/
-
-	//lua.LuaPathDefault += ";./plugins/?.lua"
-
-	err := l.DoFile("./plugins/championship.lua")
-
-	if err != nil {
-		return err, event
-	}
-
-	eventJson, err := json.Marshal(event)
-
-	if err != nil {
-		return err, event
-	}
-
-	championshipJson, err := json.Marshal(championship)
-
-	if err != nil {
-		return err, event
-	}
-
+func championshipEventStartPlugin(event *ChampionshipEvent, championship *Championship) error {
 	var standings [][]*ChampionshipStanding
 
 	for _, class := range championship.Classes {
 		standings = append(standings, class.Standings(championship.Events))
 	}
 
-	standingsJson, err := json.Marshal(standings)
+	p := &LuaPlugin{}
+
+	newEvent, newChampionship := NewChampionshipEvent(), NewChampionship(championship.Name)
+
+	p.Inputs(event, championship, standings).Outputs(newEvent, newChampionship)
+	err := p.Call("./plugins/championship.lua", "onChampionshipEventStart")
 
 	if err != nil {
-		return err, event
+		return err
 	}
 
-	if err := l.CallByParam(lua.P{
-		Fn:      l.GetGlobal("championshipEventStart"),                  // name of Lua function
-		NRet:    1,                     // number of returned values
-		Protect: true,                  // return err or panic
-	}, lua.LString(string(eventJson)), lua.LString(string(championshipJson)), lua.LString(string(standingsJson))); err != nil {
-		return err, event
-	}
+	*event, *championship = *newEvent, *newChampionship
 
-	encodedEvent := l.Get(-1)
-
-	var newEvent *ChampionshipEvent
-
-	err = json.Unmarshal([]byte(encodedEvent.String()), &newEvent)
-
-	if err != nil {
-		return err, event
-	}
-
-	return nil, newEvent
+	return nil
 }
 
 func (cm *ChampionshipManager) GetChampionshipAndEvent(championshipID string, eventID string) (*Championship, *ChampionshipEvent, error) {
@@ -711,6 +670,14 @@ func (cm *ChampionshipManager) ScheduleEvent(championshipID string, eventID stri
 			}
 		}
 
+		if config.Lua.Enabled {
+			err = championshipEventSchedulePlugin(championship, event)
+
+			if err != nil {
+				logrus.WithError(err).Error("event schedule plugin script failed")
+			}
+		}
+
 		cm.championshipEventStartTimers[event.ID.String()] = time.AfterFunc(duration, func() {
 			err := cm.StartScheduledEvent(championship, event)
 
@@ -734,6 +701,29 @@ func (cm *ChampionshipManager) ScheduleEvent(championshipID string, eventID stri
 	}
 
 	return cm.UpsertChampionship(championship)
+}
+
+func championshipEventSchedulePlugin(championship *Championship, event *ChampionshipEvent) error {
+	p := &LuaPlugin{}
+
+	var standings [][]*ChampionshipStanding
+
+	for _, class := range championship.Classes {
+		standings = append(standings, class.Standings(championship.Events))
+	}
+
+	newEvent, newChampionship := NewChampionshipEvent(), NewChampionship(championship.Name)
+
+	p.Inputs(event, championship, standings).Outputs(newEvent, newChampionship)
+	err := p.Call("./plugins/championship.lua", "onChampionshipEventSchedule")
+
+	if err != nil {
+		return err
+	}
+
+	*event, *championship = *newEvent, *newChampionship
+
+	return nil
 }
 
 func (cm *ChampionshipManager) StartScheduledEvent(championship *Championship, event *ChampionshipEvent) error {