+++ cmd/server-manager/plugins/championship.lua
@@ -1,11 +1,11 @@
 json = require "json"
 
--- these are lua hooks related to championships, for help please view luaHelp.md!
+-- these are lua hooks related to championships, for help please view lua_readme.md!
 -- there are some example functions here to give you an idea of what is possible, feel free to write your own!
 -- if you do and think other people would be interested in them consider making a pull request at https://github.com/cj123/assetto-server-manager
 
 -- called when a championship is started from the UI, before it is started on the server
-function championshipEventStart(encodedEvent, encodedChampionship, encodedClassStandings)
+function onChampionshipEventStart(encodedEvent, encodedChampionship, encodedClassStandings)
     -- Decode block, you probably shouldn't touch these!
     local event = json.decode(encodedEvent)
     local championship = json.decode(encodedChampionship)
@@ -17,14 +17,35 @@ function championshipEventStart(encodedEvent, encodedChampionship, encodedClassS
     --print("Championship:", dump(championship)) --championships can get pretty huge, this might exceed terminal limit
     --print("Standings:", dump(standings))
 
-    -- Function block
-    event = addBallastFromChampionshipPosition(event, standings)
+    -- Function block NOTE: this hook BLOCKS, make sure your functions don't loop forever!
+    -- uncomment functions to enable them!
+    --event = addBallastFromChampionshipPosition(event, standings, 50)
 
     -- Encode block, you probably shouldn't touch these either!
-    return json.encode(event)
+    return json.encode(championship), json.encode(event)
 end
 
-function addBallastFromChampionshipPosition(event, standings)
+-- called when any CHAMPIONSHIP event is scheduled
+function onChampionshipEventSchedule(encodedEvent, encodedChampionship, encodedClassStandings)
+    -- Decode block, you probably shouldn't touch these!
+    local event = json.decode(encodedEvent)
+    local championship = json.decode(encodedChampionship)
+    local standings = json.decode(encodedClassStandings)
+
+    -- Uncomment these lines and run the function (start any event) to print out the structure of each object.
+    --print("Event:", dump(event))
+    --print("Championship:", dump(championship)) --championships can get pretty huge, this might exceed terminal limit
+    --print("Standings:", dump(standings))
+
+    -- Function block NOTE: this hook BLOCKS, make sure your functions don't loop forever!
+
+
+    -- Encode block, you probably shouldn't touch these either!
+    return json.encode(championship), json.encode(event)
+end
+
+-- add ballast to drivers for the championship event based on their current championship position
+function addBallastFromChampionshipPosition(event, standings, maxBallast)
     -- loop over each championship class
     for className,classStandings in pairs(standings) do
         -- loop over the standings for the class
@@ -34,7 +55,7 @@ function addBallastFromChampionshipPosition(event, standings)
                 -- if standing and entrant guids match
                 if entrant["GUID"] == standing["Car"]["Driver"]["Guid"] then
                     -- add ballast based on championship position
-                    entrant["Ballast"] = math.floor(20/(pos))
+                    entrant["Ballast"] = math.floor(maxBallast/(pos))
                 end
             end
         end
@@ -50,7 +71,8 @@ function forceVirtualMirror(event)
     return event
 end
 
-function dump(o)
+-- @TODO you can't access this in the other files
+dump = function(o)
     if type(o) == 'table' then
         local s = '{ '
         for k,v in pairs(o) do